test: organize WTE data source tests into dedicated module
Consolidate all tests related to the World Terrestrial Ecosystems
(WTE) data source into a shared test module for improved
organization and maintainability.
