+++ tests/data_sources/test_world_terrestrial_ecosystems.py
@@ -2,10 +2,13 @@
 
 from importlib.resources import files
 import pytest
-from tests.conftest import load_geometry
+from tests.conftest import load_geometry, load_response
 from geoenvo.geometry import Geometry
 from geoenvo.data_sources import WorldTerrestrialEcosystems
-from geoenvo.data_sources.world_terrestrial_ecosystems import create_attribute_table
+from geoenvo.data_sources.world_terrestrial_ecosystems import (
+    create_attribute_table,
+    apply_code_mapping,
+)
 
 
 def test_init():
@@ -74,3 +77,28 @@ def test_attribute_tables(use_mock, tmp_path):
             ) as attribute_table:
                 fixture_content = attribute_table.read()
                 assert content == fixture_content
+
+
+def test_apply_code_mapping():
+    """Test apply_code_mapping function."""
+
+    # Positive test case - Code of a non-empty response are mapped to
+    # environmental properties
+    response = load_response("wte_success")
+    code = response.data["properties"]["Values"][0]
+    assert code == "175"
+    data = apply_code_mapping(response.data)
+    assert len(data["results"]) == 1
+    for k, v in data["results"][0].items():
+        assert isinstance(k, str)
+        assert len(k) > 0
+        assert isinstance(v, str)
+        assert len(v) > 0
+
+    # Negative test case - Codes of an empty response are not mapped to
+    # environmental properties
+    response = load_response("wte_fail")
+    code = response.data["properties"]["Values"][0]
+    assert code == "NoData"
+    data = apply_code_mapping(response.data)
+    assert data == {"results": []}
